+++ modules/em/src/FitRestraint.cpp
@@ -185,7 +185,7 @@ double FitRestraint::unprotected_evaluate(DerivativeAccumulator *accum) const {
   // To smooth the score,
   // we start considering centroids distance once 80% of the particles
   // are outside of the density.
-  escore = CoarseCC::calc_score(
+  escore = get_coarse_cc_score(
       const_cast<DensityMap *>(target_dens_map_.get()),
       const_cast<SampledDensityMap *>(model_dens_map_.get()), scalefac_, true,
       false, norm_factors_);
@@ -193,8 +193,8 @@ double FitRestraint::unprotected_evaluate(DerivativeAccumulator *accum) const {
     // calculate the derivatives for non rigid bodies
     IMP_LOG_VERBOSE("Going to calc derivatives for none_rb_model_dens_map_\n");
     const_cast<FitRestraint *>(this)->dv_ =
-        CoarseCC::calc_derivatives(target_dens_map_, model_dens_map_, all_ps_,
-                                   weight_key_, kernel_params_, scalefac_, dv_);
+        get_coarse_cc_derivatives(target_dens_map_, model_dens_map_, all_ps_,
+                                  weight_key_, kernel_params_, scalefac_, dv_);
 
     IMP_LOG_VERBOSE(
         "Finish calculating derivatives for none_rb_model_dens_map_\n");