+++ modules/em/test/medium_test_correlation.py
@@ -38,7 +38,6 @@ class Tests(IMP.test.TestCase):
         self.EM_map.std_normalize()
         self.EM_map.get_header_writable().compute_xyz_top()
 
-        self.ccc = IMP.em.CoarseCC()
         self.ccc_intervals = IMP.em.CoarseCCatIntervals()
 
     def calc_simple_correlation(self):
@@ -46,7 +45,7 @@ class Tests(IMP.test.TestCase):
         self.model_map.calcRMS()
         threshold = self.model_map.get_header().dmin - 0.1
         return (
-            self.ccc.cross_correlation_coefficient(
+            IMP.em.get_coarse_cc_coefficient(
                 self.EM_map,
                 self.model_map,
                 threshold,
@@ -63,9 +62,9 @@ class Tests(IMP.test.TestCase):
         """ test that padding option does not affect the CC score"""
         self.model_map.calcRMS()
         threshold = self.model_map.get_header().dmin - 0.001
-        score1 = self.ccc.cross_correlation_coefficient(
+        score1 = IMP.em.get_coarse_cc_coefficient(
             self.EM_map, self.model_map, threshold, False)
-        score2 = self.ccc.cross_correlation_coefficient(
+        score2 = IMP.em.get_coarse_cc_coefficient(
             self.EM_map, self.model_map, threshold, True)
         self.assertAlmostEqual(score1, score2, 2)
 
@@ -86,7 +85,7 @@ class Tests(IMP.test.TestCase):
         threshold = self.model_map.get_header().dmin
 
         self.EM_map.get_header_writable().compute_xyz_top()
-        score1 = self.ccc.cross_correlation_coefficient(
+        score1 = IMP.em.get_coarse_cc_coefficient(
             self.EM_map, self.model_map, threshold, False)
 
         # compute correlation translating the particles
@@ -95,7 +94,7 @@ class Tests(IMP.test.TestCase):
             IMP.core.XYZ(atom.get_particle()).set_coordinates(
                 translation.get_transformed(IMP.core.XYZ(atom.get_particle()).get_coordinates()))
         interval = 1
-        score = self.ccc.calc_score(self.EM_map, self.model_map, 1.0)
+        score = IMP.em.get_coarse_cc_score(self.EM_map, self.model_map, 1.0)
         score2 = 1. - score
         self.assertAlmostEqual(score1, score2, delta=.05 * (score1 + score2))
 
@@ -113,7 +112,7 @@ class Tests(IMP.test.TestCase):
 
         self.model_map.set_origin(self.xo - xm, self.yo - ym, self.zo - zm)
         interval = 1
-        score = self.ccc.calc_score(self.EM_map, self.model_map, 1.0)
+        score = IMP.em.get_coarse_cc_score(self.EM_map, self.model_map, 1.0)
         score = 1. - score
         self.assertAlmostEqual(simple_score, score, 2)
 
@@ -134,7 +133,7 @@ class Tests(IMP.test.TestCase):
         # calculate correlation
         for i in range(0, times):
             scores_wo_intervals.append(
-                self.ccc.calc_score(
+                IMP.em.get_coarse_cc_score(
                     self.EM_map,
                     self.model_map,
                     1.0))